+++ apps/meteor/app/oembed/server/jumpToMessage.ts
@@ -1,15 +1,15 @@
-import URL from 'url';
 import QueryString from 'querystring';
+import URL from 'url';
 
-import { Meteor } from 'meteor/meteor';
 import type { MessageAttachment, IMessage, IOmnichannelRoom } from '@rocket.chat/core-typings';
 import { isQuoteAttachment } from '@rocket.chat/core-typings';
 import { Messages, Users, Rooms } from '@rocket.chat/models';
+import { Meteor } from 'meteor/meteor';
 
-import { createQuoteAttachment } from '../../../lib/createQuoteAttachment';
-import { settings } from '../../settings/server';
 import { callbacks } from '../../../lib/callbacks';
+import { createQuoteAttachment } from '../../../lib/createQuoteAttachment';
 import { canAccessRoomAsync } from '../../authorization/server/functions/canAccessRoom';
+import { settings } from '../../settings/server';
 import { getUserAvatarURL } from '../../utils/server/getUserAvatarURL';
 
 const recursiveRemoveAttachments = (attachments: MessageAttachment, deep = 1, quoteChainLimit: number): MessageAttachment => {