+++ apps/meteor/client/sidebar/Item/Condensed.stories.tsx
@@ -27,7 +27,7 @@ const Template: ComponentStory<typeof Condensed> = (args) => (
 	<Condensed
 		{...args}
 		titleIcon={
-			<Box mi='x4'>
+			<Box mi={4}>
 				<Status.Online />
 			</Box>
 		}