+++ apps/meteor/jest.config.ts
@@ -3,7 +3,11 @@ export default {
 
 	testEnvironment: 'jsdom',
 	modulePathIgnorePatterns: ['<rootDir>/dist/'],
-	testMatch: ['<rootDir>/client/**/**.spec.[jt]s?(x)'],
+	testMatch: [
+		'<rootDir>/client/**/**.spec.[jt]s?(x)',
+		'<rootDir>/tests/unit/client/views/**/*.spec.{ts,tsx}',
+		'<rootDir>/tests/unit/client/providers/**/*.spec.{ts,tsx}',
+	],
 	transform: {
 		'^.+\\.(t|j)sx?$': '@swc/jest',
 	},