+++ apps/meteor/server/lib/roles/removeUserFromRoles.ts
@@ -1,6 +1,6 @@
+import { MeteorError } from '@rocket.chat/core-services';
 import type { IRole, IUser, IRoom } from '@rocket.chat/core-typings';
 import { Users, Roles } from '@rocket.chat/models';
-import { MeteorError } from '@rocket.chat/core-services';
 
 import { validateRoleList } from './validateRoleList';