+++ packages/ui-client/src/components/Card/CardFooterWrapper.tsx
@@ -1,6 +1,6 @@
 import { Box } from '@rocket.chat/fuselage';
 import type { ReactElement, ReactNode } from 'react';
 
-const CardFooterWrapper = ({ children }: { children: ReactNode }): ReactElement => <Box mbs='x16'>{children}</Box>;
+const CardFooterWrapper = ({ children }: { children: ReactNode }): ReactElement => <Box mbs={16}>{children}</Box>;
 
 export default CardFooterWrapper;