+++ ee/packages/pdf-worker/CHANGELOG.md
@@ -1,17 +1,92 @@
 # @rocket.chat/pdf-worker
 
+## 0.0.6
+
+### Patch Changes
+
+- Updated dependencies [e14ec50816]
+- Updated dependencies [9da856cc67]
+- Updated dependencies [12d97e16c2]
+- Updated dependencies [48ac55f4ea]
+  - @rocket.chat/core-typings@6.3.0
+
+## 0.0.6-rc.10
+
+### Patch Changes
+
+- @rocket.chat/core-typings@6.3.0-rc.10
+
+## 0.0.6-rc.9
+
+### Patch Changes
+
+- Updated dependencies [48ac55f4ea]
+  - @rocket.chat/core-typings@6.3.0-rc.9
+
+## 0.0.6-rc.8
+
+### Patch Changes
+
+- @rocket.chat/core-typings@6.3.0-rc.8
+
+## 0.0.6-rc.7
+
+### Patch Changes
+
+- @rocket.chat/core-typings@6.3.0-rc.7
+
+## 0.0.6-rc.6
+
+### Patch Changes
+
+- @rocket.chat/core-typings@6.3.0-rc.6
+
+## 0.0.6-rc.5
+
+### Patch Changes
+
+- @rocket.chat/core-typings@6.3.0-rc.5
+
+## 0.0.6-rc.4
+
+### Patch Changes
+
+- @rocket.chat/core-typings@6.3.0-rc.4
+
+## 0.0.6-rc.3
+
+### Patch Changes
+
+- @rocket.chat/core-typings@6.3.0-rc.3
+
+## 0.0.6-rc.2
+
+### Patch Changes
+
+- @rocket.chat/core-typings@6.3.0-rc.2
+
+## 0.0.6-rc.1
+
+### Patch Changes
+
+- @rocket.chat/core-typings@6.3.0-rc.1
+
+## 0.0.5
+
+### Patch Changes
+
+- @rocket.chat/core-typings@6.2.10
+
 ## 0.0.4
 
-### Patch Changes
-
-- @rocket.chat/core-typings@6.2.9
-
-## 0.0.3
+## 0.0.3-rc.0
 
 ### Patch Changes
 
-- Updated dependencies []:
-  - @rocket.chat/core-typings@6.2.7
+- Updated dependencies [e14ec50816]
+- Updated dependencies [9da856cc67]
+- Updated dependencies [12d97e16c2]
+  - @rocket.chat/core-typings@6.3.0-rc.0
 
 ## 0.0.2